+++ utilery/config/default.py
@@ -1,15 +1,17 @@
 DATABASES = {
-    "default": "dbname=osm user=osm password=osm host=localhost"
+    'default': {
+        'database': 'osm',
+    }
 }
 RECIPES = []
 TILEJSON = {
-    "tilejson": "2.1.0",
-    "name": "utilery",
-    "description": "A lite vector tile server",
-    "scheme": "xyz",
-    "format": "pbf",
-    "tiles": [
-        "http://vector.myserver.org/all/{z}/{x}/{y}.pbf"
+    'tilejson': '2.1.0',
+    'name': 'utilery',
+    'description': 'A lite vector tile server',
+    'scheme': 'xyz',
+    'format': 'pbf',
+    'tiles': [
+        'http://vector.myserver.org/all/{z}/{x}/{y}.pbf'
     ],
 }
 BUILTIN_PLUGINS = ['utilery.plugins.builtins.CORS']
@@ -19,4 +21,4 @@ SRID = 900913
 SCALE = 1
 BUFFER = 0
 CLIP = False
-CORS = "*"
+CORS = '*'
